Wrap docstrings
There's an issue in Lodox right now, s.t. when pandoc is installed,
any linebreaks in docstrings are rendered as `<br>`, which is not
desirable...

A quick hack for now (and since none of these docstrings have code
blocks in them), is to call `brew unlink pandoc` before `rebar3 lodox`.